Entity Operations integration removed
Building add/edit operations automatically relies on ability to build decent forms and it wasn't working with Inline Entity Form on adequate quality. All the extensions as a result were happening on the form alteration level which is quite limitting for anything.
Right now approach we have is moving us to remove Inline Entity Form module as soon as possible, and so the Entity Operations module too.